chore(EMS-3306-3308): no pdf - change your answers - working with buyer - e2e improvements
